Move all blacklist-interfacing code to bg script
- also split up existing code into different submodules in the `/src/blacklist/background` module
