+++ src/blacklist/background/constants.js
@@ -1,0 +1,11 @@
+/**
+ * Default blacklist entries.
+ */
+export const DEF_ENTRIES = ['https?://localhost', 'google.\\w+/maps']
+
+export const CONVERT_TIME_KEY = 'data-conversion-timestamp'
+
+/**
+ * Unused storage keys for old ext data
+ */
+export const OLD_EXT_MISC_KEYS = ['shouldOpenTab', 'history', 'preferences']
