+++ src/blacklist/background/interface.js
@@ -1,12 +1,5 @@
-import { remoteFunction } from 'src/util/webextensionRPC'
 import { STORAGE_KEY } from 'src/options/blacklist/constants'
-
-const dirtyEstsCache = remoteFunction('dirtyEstsCache')
-
-/**
- * Default blacklist entries.
- */
-export const defaultEntries = ['https?://localhost', 'google.\\w+/maps']
+import importStateManager from 'src/imports/background/import-state'
 
 /**
  * Given a URL and user's blacklist, checks the URL against the blacklist expressions to see if any
@@ -56,7 +49,7 @@ export async function fetchBlacklist() {
 
 async function storeBlacklist(blacklist = []) {
     const serialized = JSON.stringify(blacklist)
-    await dirtyEstsCache()
+    await importStateManager.dirtyEsts()
 
     return browser.storage.local.set({ [STORAGE_KEY]: serialized })
 }
